// src/api/v1.ts
export interface components {
    schemas: {
// ...
        SellerCredentialResponse: {
            /** Format: int64 */
            id?: number;
            /** Format: int64 */
            sellerId?: number;
            clientId?: string;
            clientSecretMasked?: string;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}